+++ project/GitHub.scala
@@ -1,0 +1,8 @@
+package akka
+
+object GitHub {
+  def url(v: String): String = {
+    val branch = if (v.endsWith("SNAPSHOT")) "master" else "v" + v
+    "http://github.com/akka/akka/tree/" + branch
+  }
+}
